+++ client/Packages/com.beamable.server/Editor/UI/Components/PublishPopup/PublishWindow.cs
@@ -15,6 +15,8 @@ namespace Beamable.Editor.Microservice.UI.Components
 {
     public class PublishWindow : CommandRunnerWindow
     {
+        private bool isSet = false;
+
         public static PublishWindow ShowPublishWindow()
         {
             var wnd = CreateInstance<PublishWindow>();
@@ -38,6 +40,17 @@ namespace Beamable.Editor.Microservice.UI.Components
         private void OnEnable()
         {
             VisualElement root = this.GetRootVisualContainer();
+
+            if (isSet)
+            {
+                this.Refresh();
+
+                Microservices.GenerateUploadModel().Then(model =>
+                {
+                    this._model = model;
+                    this.Refresh();
+                });
+            }
         }
 
         void Refresh()
@@ -65,6 +78,8 @@ namespace Beamable.Editor.Microservice.UI.Components
             e.PrepareParent();
             e.Refresh();
             Repaint();
+
+            isSet = true;
         }
     }
 }