+++ client/Packages/com.beamable/Common/Runtime/Modules/Shop/ListingContent.cs
@@ -18,7 +18,7 @@ namespace Beamable.Common.Shop
    /// #### Related Links
    /// - See the <a target="_blank" href="https://docs.beamable.com/docs/store-feature">Store</a> feature documentation
    /// - See Beamable.Api.Commerce.CommerceService script reference
-   /// 
+   ///
    /// ![img beamable-logo]
    ///
    /// </summary>
@@ -29,10 +29,10 @@ namespace Beamable.Common.Shop
    {
       [Tooltip(ContentObject.TooltipListingPrice1)]
       public ListingPrice price;
-      
+
       [Tooltip(ContentObject.TooltipListingOffer1)]
       public ListingOffer offer;
-      
+
       [Tooltip(ContentObject.TooltipOptional0 + ContentObject.TooltipActivePeriod1)]
       public OptionalPeriod activePeriod;
 
@@ -42,20 +42,20 @@ namespace Beamable.Common.Shop
 
       [Tooltip(ContentObject.TooltipOptional0 + ContentObject.TooltipCohort1 + ContentObject.TooltipRequirement2)]
       public OptionalCohort cohortRequirements;
-      
+
       [Tooltip(ContentObject.TooltipOptional0 + ContentObject.TooltipStat1 + ContentObject.TooltipRequirement2)]
       public OptionalStats playerStatRequirements;
-      
+
       [Tooltip(ContentObject.TooltipOptional0 + ContentObject.TooltipActivePeriod1 + ContentObject.TooltipRequirement2)]
       public OptionalOffers offerRequirements;
-      
+
       [Tooltip(ContentObject.TooltipOptional0 + ContentObject.TooltipActivePeriod1)]
       public OptionalSerializableDictionaryStringToString clientData;
 
       [MustBePositive]
       [Tooltip(ContentObject.TooltipOptional0 + ContentObject.TooltipDurationSeconds1 + ContentObject.TooltipActive2)]
       public OptionalInt activeDurationSeconds;
-      
+
       [Tooltip(ContentObject.TooltipOptional0 + ContentObject.TooltipDurationSeconds1 + ContentObject.TooltipActiveCooldown2)]
       [MustBePositive]
       public OptionalInt activeDurationCoolDownSeconds;
@@ -63,9 +63,12 @@ namespace Beamable.Common.Shop
       [Tooltip(ContentObject.TooltipOptional0 + ContentObject.TooltipDurationPurchaseLimit1)]
       [MustBePositive]
       public OptionalInt activeDurationPurchaseLimit;
-      
+
       [Tooltip(ContentObject.TooltipOptional0 + ContentObject.ButtonText1)]
       public OptionalString buttonText; // TODO: This is a dictionary, not a string!
+
+      [Tooltip(ContentObject.TooltipOptional0 + "schedule for when the listing will be active")]
+      public OptionalListingSchedule schedule;
 
    }
 
@@ -82,7 +85,7 @@ namespace Beamable.Common.Shop
 
       [Tooltip(ContentObject.TooltipObtainCurrency1)]
       public List<OfferObtainCurrency> obtainCurrency;
-      
+
       [Tooltip(ContentObject.TooltipObtainItem1)]
       public List<OfferObtainItem> obtainItems;
    }
@@ -175,7 +178,7 @@ namespace Beamable.Common.Shop
       [Tooltip(ContentObject.TooltipName1)]
       [CannotBeBlank]
       public string name;
-      
+
       [Tooltip(ContentObject.TooltipValue1)]
       public string value;
    }
@@ -186,11 +189,11 @@ namespace Beamable.Common.Shop
       [Tooltip(ContentObject.TooltipType1)]
       [MustBeOneOf("sku", "currency")]
       public string type;
-      
+
       [Tooltip(ContentObject.TooltipSymbol1)]
       [MustReferenceContent(false, typeof(CurrencyContent), typeof(SKUContent))]
       public string symbol;
-      
+
       [Tooltip(ContentObject.TooltipAmount1)]
       [MustBeNonNegative]
       public int amount;
@@ -214,7 +217,7 @@ namespace Beamable.Common.Shop
       [Tooltip(ContentObject.TooltipSymbol1)]
       [MustReferenceContent(AllowedTypes = new []{typeof(ListingContent)})]
       public string offerSymbol;
-      
+
       [Tooltip(ContentObject.TooltipPurchase1)]
       public OfferConstraint purchases;
    }
@@ -227,11 +230,11 @@ namespace Beamable.Common.Shop
       [Tooltip("Domain of the stat (e.g. 'platform', 'game', 'client'). Default is 'game'.")]
       [MustBeOneOf("platform", "game", "client")]
       public OptionalString domain;
-      
+
       [Tooltip("Visibility of the stat (e.g. 'private', 'public'). Default is 'private'.")]
       [MustBeOneOf("private", "public")]
       public OptionalString access;
-      
+
       [Tooltip(ContentObject.TooltipStat1)]
       [CannotBeBlank]
       public string stat;
@@ -239,11 +242,11 @@ namespace Beamable.Common.Shop
       [Tooltip(ContentObject.TooltipConstraint1)]
       [MustBeComparatorString]
       public string constraint;
-      
+
       [Tooltip(ContentObject.TooltipValue1)]
       public int value;
    }
-   
+
    [System.Serializable]
    public class CohortRequirement
    {
@@ -254,7 +257,7 @@ namespace Beamable.Common.Shop
       [Tooltip(ContentObject.TooltipCohort1)]
       [CannotBeBlank]
       public string cohort;
-      
+
       [Tooltip(ContentObject.TooltipConstraint1)]
       [MustBeComparatorString]
       public string constraint;
@@ -272,7 +275,7 @@ namespace Beamable.Common.Shop
       [Tooltip(ContentObject.TooltipConstraint1)]
       [MustBeComparatorString]
       public string constraint;
-      
+
       [Tooltip(ContentObject.TooltipValue1)]
       public int value;
    }
@@ -296,7 +299,7 @@ namespace Beamable.Common.Shop
 
    [System.Serializable]
    public class OptionalDict : Optional<ContentDictionary> { }
-   
+
    [System.Serializable]
    public class OptionalCohort : Optional<List<CohortRequirement>> {}
 }