+++ client/Packages/com.beamable/Editor/UI/Schedules/Schedule.cs
@@ -1,0 +1,17 @@
+﻿using System;
+using System.Collections.Generic;
+using Beamable.Common.Content;
+
+namespace Beamable.Editor.Schedules
+{
+    [Serializable]
+    public class ScheduleWrapper
+    {
+        public Schedule schedule;
+
+        public ScheduleWrapper(Schedule schedule)
+        {
+            this.schedule = schedule;
+        }
+    }
+}