+++ client/Packages/com.beamable/Runtime/UI/Scripts/Buss/Properties/Variable.cs
@@ -40,6 +40,7 @@ namespace Beamable.UI.Buss.Properties
 
       public VariableScope Merge(VariableScope other)
       {
+         if (other == null) return Clone();
          var next = new VariableScope
          {
             Numbers = Numbers.Merge(other.Numbers),