+++ microservice/microservice/dbmicroservice/LocalDebugService.cs
@@ -1,15 +1,10 @@
-﻿﻿using System.Net.Mime;
- using System.Net.WebSockets;
- using System.Text;
+﻿using System.Net.WebSockets;
+using System.Text;
 using System.Threading.Tasks;
-using Beamable.Common.Api.Auth;
-using Beamable.Common.Api.Content;
-using Beamable.Content;
 using EmbedIO;
- using EmbedIO.Actions;
- using EmbedIO.Routing;
-using EmbedIO.WebApi;
-using ContentService = Beamable.Server.Content.ContentService;
+using EmbedIO.Actions;
+using Swan;
+using Swan.Logging;
 
 namespace Beamable.Server {
     public class LocalDebugService {
@@ -17,22 +12,26 @@ namespace Beamable.Server {
 
         public LocalDebugService(BeamableMicroService beamableService) {
             _beamableService = beamableService;
-            var server = new WebServer(6565)
+            Swan.Logging.ConsoleLogger.Instance.LogLevel = LogLevel.Error;
+            var server = new WebServer(SharedConstants.HEALTH_PORT)
                 .WithModule(new ActionModule("/health", HttpVerbs.Any, HealthCheck));
             server.RunAsync();
         }
 
         private async Task HealthCheck(IHttpContext context) {
             var healthy = _beamableService.HasInitialized;
-            healthy &= _beamableService.GetWebsocketPromise().IsCompleted;
+            healthy &= _beamableService.GetWebsocketPromise()?.IsCompleted ?? false;
             if (healthy && !_beamableService.IsShuttingDown) {
                 var connection = _beamableService.GetWebsocketPromise().GetResult();
                 healthy &= connection.State == WebSocketState.Open;
             }
             if (healthy) {
+                context.SetHandled();
                 await context.SendStringAsync("Healthy: true", "text", Encoding.Default);
             }
-            else {
+            else
+            {
+                context.Response.StatusCode = 409;
                 await context.SendStringAsync("Healthy: false", "text", Encoding.Default);
             }
         }